Enhanced Cassandra health checks
Motivation:

The current Cassandra health check simply attempts to query one
node in the cluster, randomly chosen, and sets the status to
DOWN if that node cannot be queried.

This is admittedly misleading, since Cassandra clusters are usually
resilient to the loss of many nodes, and can still continue to
serve application queries even if many nodes are down.

Modification:

This commit replaces the current health check by the DataStax
driver's SessionDiagnostics feature, which perform the following
checks:

- Topology: reports how many nodes were found in each datacenter
and globally, and how many were up or down.
- Token range availability: inspects the ring and detects if
certain token ranges are unavailable for the configured
keyspace, local datacenter and consistency level.

Note that these checks do not require querying the cluster; they
rely solely on metadata exposed by the driver and on Gossip events,
and thus execute very fast.

The status is set to DOWN if:

1. The entire cluster is down; or
2. There are unavailable token ranges.

Result:

The health report now accurately describes the cluster health
by setting it status to DOWN when queries are likely to
fail due to UnavailableException errors.